+++ src/core/request/interface.ts
@@ -587,9 +587,7 @@ export interface RetryOptions<D = unknown> {
 	delay?(attempt: number, error: RequestError<D>): number | Promise<void> | false;
 }
 
-export type RequestAPIValue<T = string> = Nullable<T> | ((params: RequestAPIParams) => Nullable<T>);
-
-export type RequestAPIParams = MiddlewareParams;
+export type RequestAPIValue<T = string> = Nullable<T> | ((params: MiddlewareParams) => Nullable<T>);
 
 /**
  * A map of API parameters.