+++ CHANGELOG.md
@@ -17,11 +17,18 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Now `MiddlewareParams` are being passed to the `RequestAPIValue` function `core/request`
 
-# v4.0.0-alpha.?? (2024-??-??)
+## v4.0.0-alpha.29 (2024-04-04)
+
+#### :boom: Breaking Change
+
+* Dropped API support with rest parameters in the `addToPrototype` `core/prelude/function/extend`.
+  Migration Guide: `Foo.addToPrototype(bar, baz)` -> `Foo.addToPrototype({bar, baz})`.
 
 #### :bug: Bug Fix
 
 * Fixed `providerOptions` inheritance by ExtraProviders `core/data`
+* Fixed the bug with the `debounce` decorator:
+  conflicts between debounced methods across instances of identical classes or components have been resolved `core/functools`
 
 ## v4.0.0-alpha.28 (2024-03-21)